requirements.txt
pillow
tk
tkinterdnd2